netusb: Refactor ECM code
Move ECM functionality from netusb to function_ecm making defines not
needed in many cases.
